Move logging connection classes to another module -why don't people just use a logging proxy instead? so much easier!
